images: Disable cache when using yum
And move the installation of software into the shared image. This causes our
shared image to be slightly bloated (nginx and php-fpm, and a lot of perl),
which is unnecessary for the submit image, but causes less disk io in general.

For future these two might be split out into it's own image that admin and
frontend can use.