python3-image: Drop pandas again
Pandas causes a rather large dep-tree on f31 due to pulling in a compilation
environment for fortran77, including Perl, git, gcc, and development headers.

This seems to be adjusted in f32 but it's still a hefty price to pay.

With the changes in ModioAB/afase!355  we don't need pandas in our current
generic python environment either.

Thus, drop it.

Size savings: approx 375M