Sync the rust and rust-cross containers
This gets rid of debian9 container, and makes a debian10 rust-cross
container for our use.

This container is mostly the same as our rust container, except the main
rust container supports the latest debian, suitable to build for servers
and containers, while this uses the same debian base as our devices.

If we stay on top of things, then we should always be able to use the
rust container, but that seems unlikely.