rust: Add cargo-llvm-cov tool as well
This is a tool that can both generate _some_ reports (not cobertura atm
though) but also sets up cargo with environment variables and build to
run llvm profiling.

It can also report to the terminal, which unfortunately is needed for
the _badges_ and the % coverage reports in the MR UI in gitlab, even if
we have a perfectly valid XML report that contains all that data.