Define container build template locally
Extending the `.container:build` template got weird now that
`/ci/container.yml` has its own generalized instance of the 
`.container:build-publish` template.

As this worked out, GitLab CI merged our version of the template on
top of the imported one. This means that there are rules that already
change `IMAGE_TAG_SUFFIX` to `latest` when we're on the default branch,
so we actually pushed the `latest` tag twice, instead of pushing the
branch tag _and_ the `latest` tag, as originally intended.